# api/openapi.yaml
openapi: "3.1.0"
info:
  title: OpenClarion API
  description: |
    Natural-language-driven alert governance platform.
    This specification is the single source of truth for all API contracts.
  version: 0.1.0
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/config/notification-channels:
    get:
      operationId: listNotificationChannelProfiles
      summary: List notification channel profiles
      description: Returns operator-managed notification channel profiles without endpoint URLs or credential values.
      tags:
        - config
      parameters:
        - $ref: "#/components/parameters/ListLimit"
      responses:
        "200":
          description: Notification channel profiles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationChannelProfileListResponse"
              examples:
                channels:
                  value:
                    items:
                      - id: 1
                        name: Operations webhook
                        kind: webhook
                        secret_ref: secret/example/ops-webhook
                        delivery_scopes:
                          - report
                          - diagnosis_close
                        enabled: false
                        labels:
                          team: ops
                        created_at: "2026-06-05T09:00:00Z"
                        updated_at: "2026-06-05T09:00:00Z"
        "400":
          description: Notification channel list parameters are invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid-limit:
                  value:
                    error: "limit must be between 1 and 500"
        "500":
          description: Notification channel list failed server-side
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server-error:
                  value:
                    error: list notification channel profiles failed
    post:
      operationId: createNotificationChannelProfile
      summary: Create a notification channel profile
      description: Stores notification channel metadata with a deployment-managed secret reference only. This operation does not send notifications.
      tags:
        - config
      requestBody:
        description: Notification channel profile metadata.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationChannelProfileWriteRequest"
            examples:
              webhook:
                value:
                  name: Operations webhook
                  kind: webhook
                  secret_ref: secret/example/ops-webhook
                  delivery_scopes:
                    - report
                    - diagnosis_close
                  enabled: false
                  labels:
                    team: ops
      responses:
        "201":
          description: Notification channel profile created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationChannelProfile"
              examples:
                created:
                  value:
                    id: 1
                    name: Operations webhook
                    kind: webhook
                    secret_ref: secret/example/ops-webhook
                    delivery_scopes:
                      - report
                      - diagnosis_close
                    enabled: false
                    labels:
                      team: ops
                    created_at: "2026-06-05T09:00:00Z"
                    updated_at: "2026-06-05T09:00:00Z"
        "400":
          description: Notification channel profile request is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid:
                  value:
                    error: "notification channel profile: secret_ref must be non-empty"
        "409":
          description: Notification channel profile conflicts with an existing profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                duplicate:
                  value:
                    error: notification channel profile already exists
        "500":
          description: Notification channel profile creation failed server-side
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server-error:
                  value:
                    error: create notification channel profile failed
  /api/v1/config/notification-channels/{channel_id}:
    get:
      operationId: getNotificationChannelProfile
      summary: Get a notification channel profile
      description: Returns one operator-managed notification channel profile without endpoint URLs or credential values.
      tags:
        - config
      parameters:
        - $ref: "#/components/parameters/NotificationChannelProfileID"
      responses:
        "200":
          description: Notification channel profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationChannelProfile"
              examples:
                channel:
                  value:
                    id: 1
                    name: Operations webhook
                    kind: webhook
                    secret_ref: secret/example/ops-webhook
                    delivery_scopes:
                      - report
                      - diagnosis_close
                    enabled: true
                    labels:
                      team: ops
                    created_at: "2026-06-05T09:00:00Z"
                    updated_at: "2026-06-05T09:05:00Z"
        "404":
          description: Notification channel profile was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                not-found:
                  value:
                    error: notification channel profile not found
        "500":
          description: Notification channel profile lookup failed server-side
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server-error:
                  value:
                    error: get notification channel profile failed
    put:
      operationId: replaceNotificationChannelProfile
      summary: Replace a notification channel profile
      description: Replaces mutable notification channel profile metadata. This operation does not send notifications.
      tags:
        - config
      parameters:
        - $ref: "#/components/parameters/NotificationChannelProfileID"
      requestBody:
        description: Replacement notification channel profile metadata.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationChannelProfileWriteRequest"
            examples:
              disable-channel:
                value:
                  name: Operations webhook
                  kind: webhook
                  secret_ref: secret/example/ops-webhook
                  delivery_scopes:
                    - report
                  enabled: false
                  labels:
                    team: ops
      responses:
        "200":
          description: Notification channel profile replaced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationChannelProfile"
              examples:
                replaced:
                  value:
                    id: 1
                    name: Operations webhook
                    kind: webhook
                    secret_ref: secret/example/ops-webhook
                    delivery_scopes:
                      - report
                    enabled: false
                    labels:
                      team: ops
                    created_at: "2026-06-05T09:00:00Z"
                    updated_at: "2026-06-05T09:10:00Z"
        "400":
          description: Notification channel profile replacement request is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid:
                  value:
                    error: "notification channel profile: delivery_scopes must be non-empty"
        "404":
          description: Replacement target profile was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                not-found:
                  value:
                    error: notification channel profile not found
        "409":
          description: Replacement profile conflicts with an existing name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                duplicate:
                  value:
                    error: notification channel profile already exists
        "500":
          description: Notification channel profile replacement failed server-side
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server-error:
                  value:
                    error: replace notification channel profile failed
  /api/v1/config/notification-channels/{channel_id}/test:
    post:
      operationId: testNotificationChannelProfile
      summary: Test a notification channel profile
      description: Sends one bounded backend test notification for a persisted notification channel profile and returns sanitized status only. This action does not create report delivery records.
      tags:
        - config
      parameters:
        - $ref: "#/components/parameters/NotificationChannelProfileID"
      responses:
        "200":
          description: Sanitized notification channel test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationChannelTestResult"
              examples:
                success:
                  value:
                    channel_id: 1
                    kind: webhook
                    status: success
                    reason_code: ok
                    message: Notification channel test delivery succeeded.
                    checked_at: "2026-06-05T10:00:00Z"
                    provider_message_id: msg-42
                    provider_status: accepted
                blocked:
                  value:
                    channel_id: 2
                    kind: webhook
                    status: blocked
                    reason_code: credentials_unavailable
                    message: Secret-backed notification channel tests require a server-side secret resolver.
                    checked_at: "2026-06-05T10:00:00Z"
                    provider_message_id: ""
                    provider_status: ""
        "400":
          description: Notification channel test request is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid:
                  value:
                    error: channel_id must be positive
        "404":
          description: Notification channel profile was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                not-found:
                  value:
                    error: notification channel profile not found
        "503":
          description: Notification channel tester is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unconfigured:
                  value:
                    error: notification channel tester is not configured
        "500":
          description: Notification channel test failed server-side
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server-error:
                  value:
                    error: test notification channel failed
...
components:
  parameters:
    ListLimit:
      name: limit
      in: query
      description: Maximum number of rows to return.
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 500
        default: 100
...
    NotificationChannelProfileID:
      name: channel_id
      in: path
      description: Notification channel profile identifier.
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
  schemas:
...
    NotificationChannelKind:
      type: string
      description: Supported notification channel adapter kind.
      enum:
        - webhook
      examples:
        - webhook
    NotificationDeliveryScope:
      type: string
      description: Notification flow that may use a channel once workflow binding is implemented.
      enum:
        - report
        - diagnosis_close
      examples:
        - report
    NotificationChannelTestStatus:
      type: string
      description: Sanitized notification-channel test outcome category.
      enum:
        - success
        - failed
        - unsupported
        - blocked
      examples:
        - success
    NotificationChannelTestReasonCode:
      type: string
      description: Stable machine-readable reason for the notification-channel test outcome.
      enum:
        - ok
        - unsupported_kind
        - credentials_unavailable
        - provider_unreachable
        - provider_error
        - invalid_profile
      examples:
        - ok
    NotificationChannelLabels:
      type: object
      description: Operator labels for notification ownership, environment, or routing metadata.
      additionalProperties:
        type: string
      examples:
        - team: ops
          env: test
    NotificationChannelProfile:
      type: object
      description: Operator-managed notification channel profile without endpoint URLs or credential values.
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          examples:
            - 1
        name:
          type: string
          minLength: 1
          maxLength: 120
          examples:
            - Operations webhook
        kind:
          $ref: "#/components/schemas/NotificationChannelKind"
        secret_ref:
          type: string
          minLength: 1
          maxLength: 256
          description: Deployment-managed endpoint secret reference; never the endpoint URL or credential value.
          examples:
            - secret/example/ops-webhook
        delivery_scopes:
          type: array
          minItems: 1
          maxItems: 2
          description: Notification flows that may use this channel after later workflow binding.
          items:
            $ref: "#/components/schemas/NotificationDeliveryScope"
          examples:
            - - report
              - diagnosis_close
        enabled:
          type: boolean
          description: Whether operators enabled this channel for future notification binding.
          examples:
            - false
        labels:
          $ref: "#/components/schemas/NotificationChannelLabels"
        created_at:
          type: string
          format: date-time
          examples:
            - "2026-06-05T09:00:00Z"
        updated_at:
          type: string
          format: date-time
          examples:
            - "2026-06-05T09:10:00Z"
      required:
        - id
        - name
        - kind
        - secret_ref
        - delivery_scopes
        - enabled
        - labels
        - created_at
        - updated_at
    NotificationChannelProfileWriteRequest:
      type: object
      description: Notification channel profile metadata accepted by create and replace operations.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
          examples:
            - Operations webhook
        kind:
          $ref: "#/components/schemas/NotificationChannelKind"
        secret_ref:
          type: string
          minLength: 1
          maxLength: 256
          description: Submitted deployment-managed endpoint secret reference; never the endpoint URL or credential value.
          examples:
            - secret/example/ops-webhook
        delivery_scopes:
          type: array
          minItems: 1
          maxItems: 2
          items:
            $ref: "#/components/schemas/NotificationDeliveryScope"
          examples:
            - - report
              - diagnosis_close
        enabled:
          type: boolean
          default: false
        labels:
          $ref: "#/components/schemas/NotificationChannelLabels"
      required:
        - name
        - kind
        - secret_ref
        - delivery_scopes
      additionalProperties: false
    NotificationChannelTestResult:
      type: object
      description: Sanitized result for an explicit notification-channel delivery test.
      properties:
        channel_id:
          type: integer
          format: int64
          minimum: 1
          examples:
            - 1
        kind:
          $ref: "#/components/schemas/NotificationChannelKind"
        status:
          $ref: "#/components/schemas/NotificationChannelTestStatus"
        reason_code:
          $ref: "#/components/schemas/NotificationChannelTestReasonCode"
        message:
          type: string
          minLength: 1
          maxLength: 240
          description: Operator-facing sanitized message. It must not include endpoint URLs, secret references, tokens, or raw provider error text.
          examples:
            - Notification channel test delivery succeeded.
        checked_at:
          type: string
          format: date-time
          examples:
            - "2026-06-05T10:00:00Z"
        provider_message_id:
          type: string
          maxLength: 128
          description: Optional bounded provider acknowledgement identifier, when available from a successful provider response.
          examples:
            - msg-42
        provider_status:
          type: string
          maxLength: 64
          description: Optional bounded provider acknowledgement status, when available from a successful provider response.
          examples:
            - accepted
      required:
        - channel_id
        - kind
        - status
        - reason_code
        - message
        - checked_at
        - provider_message_id
        - provider_status
    NotificationChannelProfileListResponse:
      type: object
      description: Notification channel profile list response.
      properties:
        items:
          type: array
          examples:
            -
              - id: 1
                name: Operations webhook
                kind: webhook
                secret_ref: secret/example/ops-webhook
                delivery_scopes:
                  - report
                  - diagnosis_close
                enabled: false
                labels:
                  team: ops
                created_at: "2026-06-05T09:00:00Z"
                updated_at: "2026-06-05T09:00:00Z"
          items:
            $ref: "#/components/schemas/NotificationChannelProfile"
      required:
        - items
...
    ErrorResponse:
      type: object
      description: Error response payload.
      properties:
        error:
          type: string
          examples:
            - "limit must be between 1 and 500 (got 0)"
      required:
        - error
      examples:
        - error: "limit must be between 1 and 500 (got 0)"